all but done with crd for second feature
